integrate auth_views to views.py
